+++ changelog.md
@@ -2,6 +2,7 @@
 Scenery Changelog
 =================
 
+* 2013-5-30 Font is now immutable, use font.with( { ... } ) to create a modified copy
 * 2013-5-21 Touches now trigger the enter/over and out/exit before/after the 'down' and 'up'/'cancel' events, respectively
 * 2013-5-21 Added ButtonListener
 * 2013-5-21 Node.bounds changed to behavior of Node.visibleBounds, Node.completeBounds now handles the old Node.bounds behavior